[kernel] quick paging structure fixes & test

+++ kernel/include/paging.h
@@ -17,16 +17,16 @@
 
 typedef struct PACKED paging_pml4_entry {
   /* 8 bytes */
-  unsigned int present         : 1;
-  unsigned int writable        : 1;
-  unsigned int user            : 1;  // accessible to user-mode if 1
-  unsigned int write_through   : 1;
-  unsigned int cache_disable   : 1;
-  unsigned int accessed        : 1;
-  unsigned int zero1           : 6;
-  unsigned int pdpt_physical   : 40; // shift 12
-  unsigned int zero2           : 11;
-  unsigned int execute_disable : 1;
+  unsigned long present         : 1;
+  unsigned long writable        : 1;
+  unsigned long user            : 1;  // accessible to user-mode if 1
+  unsigned long write_through   : 1;
+  unsigned long cache_disable   : 1;
+  unsigned long accessed        : 1;
+  unsigned long zero1           : 6;
+  unsigned long pdpt_physical   : 40; // shift 12
+  unsigned long zero2           : 11;
+  unsigned long execute_disable : 1;
 } paging_pml4_entry_t;
 
 typedef paging_pml4_entry_t paging_pml4_t[PAGING_PML4_SIZE];
@@ -37,36 +37,36 @@ typedef paging_pml4_entry_t paging_pml4_t[PAGING_PML4_SIZE];
 
 typedef struct PACKED paging_pdpt_pointer_entry {
   /* 8 bytes */
-  unsigned int present         : 1;
-  unsigned int writable        : 1;
-  unsigned int user            : 1;  // accessible to user-mode if 1
-  unsigned int write_through   : 1;
-  unsigned int cache_disable   : 1;
-  unsigned int accessed        : 1;
-  unsigned int zero1           : 6;
-  unsigned int pd_physical     : 40; // shift 12
-  unsigned int zero2           : 11;
-  unsigned int execute_disable : 1;
+  unsigned long present         : 1;
+  unsigned long writable        : 1;
+  unsigned long user            : 1;  // accessible to user-mode if 1
+  unsigned long write_through   : 1;
+  unsigned long cache_disable   : 1;
+  unsigned long accessed        : 1;
+  unsigned long zero1           : 6;
+  unsigned long pd_physical     : 40; // shift 12
+  unsigned long zero2           : 11;
+  unsigned long execute_disable : 1;
 } paging_pdpt_pointer_entry_t;
 
 /* For 1 GB pages */
 typedef struct PACKED paging_pdpt_page_entry {
   /* 8 bytes */
-  unsigned int present         : 1;
-  unsigned int writable        : 1;
-  unsigned int user            : 1;
-  unsigned int write_through   : 1;
-  unsigned int cache_disable   : 1;
-  unsigned int accessed        : 1;
-  unsigned int dirty           : 1;
-  unsigned int page_size       : 1;  // must be 1
-  unsigned int global          : 1;
-  unsigned int zero1           : 3;
-  unsigned int pat             : 1;  // affects caching somehow
-  unsigned int zero2           : 17;
-  unsigned int page_physical   : 22; // shift 30
-  unsigned int zero3           : 11;
-  unsigned int execute_disable : 1;
+  unsigned long present         : 1;
+  unsigned long writable        : 1;
+  unsigned long user            : 1;
+  unsigned long write_through   : 1;
+  unsigned long cache_disable   : 1;
+  unsigned long accessed        : 1;
+  unsigned long dirty           : 1;
+  unsigned long page_size       : 1;  // must be 1
+  unsigned long global          : 1;
+  unsigned long zero1           : 3;
+  unsigned long pat             : 1;  // affects caching somehow
+  unsigned long zero2           : 17;
+  unsigned long page_physical   : 22; // shift 30
+  unsigned long zero3           : 11;
+  unsigned long execute_disable : 1;
 } paging_pdpt_page_entry_t;
 
 typedef union paging_pdpt_entry {
@@ -75,8 +75,8 @@ typedef union paging_pdpt_entry {
   paging_pdpt_page_entry_t    as_page;
 
   struct {
-    unsigned int ignore1   : 7;
-    unsigned int page_size : 1; // 0 => as_pointer, 1 => as_page
+    unsigned long ignore1   : 7;
+    unsigned long page_size : 1; // 0 => as_pointer, 1 => as_page
   } info;
 } paging_pdpt_entry_t;
 
@@ -88,36 +88,36 @@ typedef paging_pdpt_entry_t paging_pdpt_t[PAGING_PDPT_SIZE];
 
 typedef struct PACKED paging_pd_pointer_entry {
   /* 8 bytes */
-  unsigned int present         : 1;
-  unsigned int writable        : 1;
-  unsigned int user            : 1;  // accessible to user-mode if 1
-  unsigned int write_through   : 1;
-  unsigned int cache_disable   : 1;
-  unsigned int accessed        : 1;
-  unsigned int zero1           : 6;
-  unsigned int pt_physical     : 40; // shift 12
-  unsigned int zero2           : 11;
-  unsigned int execute_disable : 1;
+  unsigned long present         : 1;
+  unsigned long writable        : 1;
+  unsigned long user            : 1;  // accessible to user-mode if 1
+  unsigned long write_through   : 1;
+  unsigned long cache_disable   : 1;
+  unsigned long accessed        : 1;
+  unsigned long zero1           : 6;
+  unsigned long pt_physical     : 40; // shift 12
+  unsigned long zero2           : 11;
+  unsigned long execute_disable : 1;
 } paging_pd_pointer_entry_t;
 
 /* For 2 MB pages */
 typedef struct PACKED paging_pd_page_entry {
   /* 8 bytes */
-  unsigned int present         : 1;
-  unsigned int writable        : 1;
-  unsigned int user            : 1;
-  unsigned int write_through   : 1;
-  unsigned int cache_disable   : 1;
-  unsigned int accessed        : 1;
-  unsigned int dirty           : 1;
-  unsigned int page_size       : 1;  // must be 1
-  unsigned int global          : 1;
-  unsigned int zero1           : 3;
-  unsigned int pat             : 1;  // affects caching somehow
-  unsigned int zero2           : 8;
-  unsigned int page_physical   : 31; // shift 21
-  unsigned int zero3           : 11;
-  unsigned int execute_disable : 1;
+  unsigned long present         : 1;
+  unsigned long writable        : 1;
+  unsigned long user            : 1;
+  unsigned long write_through   : 1;
+  unsigned long cache_disable   : 1;
+  unsigned long accessed        : 1;
+  unsigned long dirty           : 1;
+  unsigned long page_size       : 1;  // must be 1
+  unsigned long global          : 1;
+  unsigned long zero1           : 3;
+  unsigned long pat             : 1;  // affects caching somehow
+  unsigned long zero2           : 8;
+  unsigned long page_physical   : 31; // shift 21
+  unsigned long zero3           : 11;
+  unsigned long execute_disable : 1;
   /* 8 bytes */
 } paging_pd_page_entry_t;
 
@@ -127,8 +127,8 @@ typedef union paging_pd_entry {
   paging_pd_page_entry_t    as_page;
 
   struct {
-    unsigned int ignore1   : 7;
-    unsigned int page_size : 1; // 0 => as_pointer, 1 => as_page
+    unsigned long ignore1   : 7;
+    unsigned long page_size : 1; // 0 => as_pointer, 1 => as_page
   } info;
 } paging_pd_entry_t;
 
@@ -141,17 +141,17 @@ typedef paging_pd_entry_t paging_pd_t[PAGING_PD_SIZE];
 /* For 4 KB pages */
 typedef struct PACKED paging_pt_entry {
   /* 8 bytes */
-  unsigned int present         : 1;
-  unsigned int writable        : 1;
-  unsigned int user            : 1;
-  unsigned int write_through   : 1;
-  unsigned int cache_disable   : 1;
-  unsigned int accessed        : 1;
-  unsigned int dirty           : 1;
-  unsigned int pat             : 1;  // affects caching somehow
-  unsigned int global          : 1;
-  unsigned int zero1           : 3;
-  unsigned int page_physical   : 40; // shift 12
-  unsigned int zero2           : 11;
-  unsigned int execute_disable : 1;
+  unsigned long present         : 1;
+  unsigned long writable        : 1;
+  unsigned long user            : 1;
+  unsigned long write_through   : 1;
+  unsigned long cache_disable   : 1;
+  unsigned long accessed        : 1;
+  unsigned long dirty           : 1;
+  unsigned long pat             : 1;  // affects caching somehow
+  unsigned long global          : 1;
+  unsigned long zero1           : 3;
+  unsigned long page_physical   : 40; // shift 12
+  unsigned long zero2           : 11;
+  unsigned long execute_disable : 1;
 } paging_pt_entry_t;
